Refactor la structure des composants de la table des personnes et ajoute des filtres et une pagination
