+++ src/components/persons/table-pagination.tsx
@@ -1,0 +1,34 @@
+import { Button } from "@/components/ui/button";
+import { Person } from "@prisma/client";
+import { Table as ReactTable } from "@tanstack/react-table";
+
+export const PersonsTablePagination = ({
+  table,
+}: {
+  table: ReactTable<Person>;
+}) => (
+  <div className="flex items-center justify-end space-x-2 py-4">
+    <div className="flex-1 text-sm text-muted-foreground">
+      {table.getFilteredSelectedRowModel().rows.length} of{" "}
+      {table.getFilteredRowModel().rows.length} row(s) selected.
+    </div>
+    <div className="space-x-2">
+      <Button
+        variant="outline"
+        size="sm"
+        onClick={() => table.previousPage()}
+        disabled={!table.getCanPreviousPage()}
+      >
+        Previous
+      </Button>
+      <Button
+        variant="outline"
+        size="sm"
+        onClick={() => table.nextPage()}
+        disabled={!table.getCanNextPage()}
+      >
+        Next
+      </Button>
+    </div>
+  </div>
+);
